task: added component for new field form
